requirements.txt
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.6
iniconfig==2.0.0
packaging==24.0
pdfminer.six==20231228
pillow==10.3.0
pluggy==1.4.0
pycparser==2.22
pyparsing==3.1.2
pypdf==4.1.0
pytest==8.1.1
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
sortedcontainers==2.4.0
toml==0.10.2
urllib3==2.2.1
